Encryption_App (2)/Encryption_App/requirements.txt
Flask==2.0.1
pycryptodome==3.14.1
cryptography==40.0.0
